Integração do Timer e Correção de bugs

+++ src/com/pomopet/data/Animal.java
@@ -2,6 +2,6 @@ package com.pomopet.data;
 
 public interface Animal {
  
-    public void calculateGrow(int xp);
+    public void calculateGrow(double xp);
     
 }
